rename package metric to metrics
