+++ pkg/metrics/serve.go
@@ -18,7 +18,7 @@ import (
 	"github.com/containerd/containerd/log"
 	"github.com/containerd/nydus-snapshotter/pkg/daemon"
 	"github.com/containerd/nydus-snapshotter/pkg/manager"
-	"github.com/containerd/nydus-snapshotter/pkg/metric/exporter"
+	"github.com/containerd/nydus-snapshotter/pkg/metrics/exporter"
 )
 
 type ServerOpt func(*Server) error
